requirements.txt
# All required runtime dependencies (GPU mandatory) for ASR + TTS + separation

# --- GPU runtime prerequisite (install first) ---
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu128

# Core numerical/audio
numpy==1.26.2
scipy==1.15.3
soundfile==0.13.1
soxr==1.0.0
librosa==0.10.2.post1
pydub==0.25.1
pyloudnorm==0.1.1
audioread==3.0.1

# Deep Learning Framework
accelerate==1.8.1
pytorch-lightning==2.5.5

# ASR stack
ctranslate2==4.6.0
faster-whisper==1.2.0
sentencepiece==0.2.1
funasr==1.0.27
modelscope==1.27.0
pyannote.audio==3.4.0

# Hugging Face ecosystem
huggingface-hub==0.35.1
safetensors==0.5.2
transformers==4.52.1
tokenizers==0.21.0

# TTS engines
auraloss==0.4.0
pesq==0.0.4
pystoi==0.4.1
einops==0.6.1
rotary-embedding-torch==0.3.5
x-transformers

# Text processing for TTS
g2p-en==2.1.0
jieba==0.42.1
jieba-fast
cn2an==0.5.22
textstat==0.7.10
pyphen==0.17.2
inflect==7.5.0
pypinyin
g2pk2
ko-pron
split-lang
fast-langdetect>=0.3.1
pyopenjtalk>=0.4.1
peft
ToJyutping
wordsegment
opencc-python-reimplemented
chardet
# WeTextProcessing==1.0.4.1  # Requires pynini, use precompiled wheel on Windows
# Windows: pip install pynini-2.1.6.post1-cp310-cp310-win_amd64.whl
# Then: pip install WeTextProcessing==1.0.4.1

# Audio/vocal separation (MSST-WebUI)
ml_collections==1.1.0
beartype==0.14.1
demucs==4.0.0
openunmix==1.3.0
asteroid==0.7.0
asteroid-filterbanks==0.4.0
descript-audiotools==0.7.2
pedalboard==0.8.9
torch-pitch_shift==1.2.5
samplerate==0.2.1
spafe==0.3.2
omegaconf==2.3.0
prodigyopt==1.1.2
sageattention==1.0.6

# Audio processing and models
timm==0.9.2
efficientnet_pytorch==0.7.1
segmentation-models-pytorch==0.3.3
torchseg==0.0.1a1
nnAudio
ninja==1.13.0

# Subtitles and timing
pysubs2==1.8.0
srt==3.5.3
regex==2025.9.18

# Media I/O
ffmpeg-python==0.2.0
yt-dlp
#you-get(备用)
av==15.1.0
opencv-python==4.9.0.80

# ML utilities
scikit-learn==1.7.2
joblib==1.5.2

# API and Web
fastapi==0.117.1
uvicorn==0.37.0
gradio==5.47.2
gradio_client==1.13.3
python-multipart==0.0.20

# Utilities
requests==2.32.5
httpx==0.28.1
aiohttp==3.12.15
orjson==3.11.3
python-dotenv==1.1.1
tqdm==4.67.1
rich==14.1.0
psutil==7.1.0
fire==0.7.1
click==8.3.0
coloredlogs==15.0.1

# Development tools
ipython==8.37.0

